airflow/requirements.txt
apache-airflow-providers-google
pyarrow
pandas
sqlalchemy
psycopg2-binary
requests
google-cloud-bigquery-storage
selenium
webdriver-manager